requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
motor==3.3.2
PyMuPDF==1.23.14
sentence-transformers==2.2.2
numpy==1.24.3
faiss-cpu==1.7.4
pydantic==2.5.0
